Create README.md
SONTEC - fotowoltaika, pompy ciepła, magazyny energii, klimatyzacja, rekuperacja, akumulatory energii

In the process of developing a website for a service company, utilizing HTML, CSS, JavaScript, and React to focus on intuitive design and functionality that enhances user experience. This project is an opportunity to apply modern web development practices.